+++ jena-text-es/src/main/java/org/apache/jena/query/text/es/ESSettings.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.jena.query.text;
+package org.apache.jena.query.text.es;
 
 import java.util.HashMap;
 import java.util.Map;
