+++ jena-text-es/src/main/resources/META-INF/NOTICE
@@ -1,0 +1,5 @@
+Apache Jena - jena-text module
+Copyright 2011-2017 The Apache Software Foundation
+
+This product includes software developed at
+The Apache Software Foundation (http://www.apache.org/).
