+++ jena-text-es/src/main/resources/META-INF/services/org.apache.jena.system.JenaSubsystemLifecycle
@@ -1,0 +1,1 @@
+org.apache.jena.query.text.es.InitJenaTextES
