+++ jena-text-es/src/main/resources/org/apache/jena/query/text/properties.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE properties SYSTEM "http://java.sun.com/dtd/properties.dtd">
+<!--  Licensed under the terms of http://www.apache.org/licenses/LICENSE-2.0 -->
+<properties version="1.0">
+  <comment>jena-text system properties</comment>
+  <entry key="org.apache.jena.query.text.version">${project.version}</entry>
+  <entry key="org.apache.jena.query.text.datetime">${build.time.xsd}</entry>
+</properties>
+
