+++ jena-text-es/testing/TextQuery/data.skos
@@ -1,0 +1,36 @@
+﻿<?xml version="1.0" encoding="UTF-8"?>
+<rdf:RDF
+    xmlns:rdf="http://www.w3.org/1999/02/22-rdf-syntax-ns#"
+    xmlns:rdfs="http://www.w3.org/2000/01/rdf-schema#"
+    xmlns:skos="http://www.w3.org/2004/02/skos/core#"
+    xmlns:dc="http://purl.org/dc/elements/1.1/">
+
+    <skos:ConceptScheme rdf:about="http://example.com/dishes">
+        <dc:title>The Example Taxonomy</dc:title>
+        <dc:description>An example taxonomy to illustrate the use of the SKOS schema.</dc:description>
+    </skos:ConceptScheme>
+
+    <skos:Concept rdf:about="http://example.com/dishes#potatoBased">
+        <skos:prefLabel xml:lang="fr">Plats à base de pomme de terre</skos:prefLabel>
+        <skos:prefLabel xml:lang="en">Potato based dishes</skos:prefLabel>
+        <skos:prefLabel xml:lang="de">Kartoffelgerichte</skos:prefLabel>
+        <skos:inScheme rdf:resource="http://example.com/dishes"/>
+        <skos:topConceptOf rdf:resource="http://example.com/dishes"/>
+    </skos:Concept>
+
+    <skos:Concept rdf:about="http://example.com/dishes#fries">
+        <skos:prefLabel xml:lang="fr">Frites</skos:prefLabel>
+        <skos:prefLabel xml:lang="en">French fries</skos:prefLabel>
+        <skos:prefLabel xml:lang="de">Französisch frites</skos:prefLabel>
+        <skos:inScheme rdf:resource="http://example.com/dishes"/>
+        <skos:broader rdf:resource="http://example.com/dishes#potatoBased"/>
+    </skos:Concept>
+
+    <skos:Concept rdf:about="http://example.com/dishes#mashed">
+        <skos:prefLabel xml:lang="fr">Purée de pomme de terre</skos:prefLabel>
+        <skos:prefLabel xml:lang="en">Mashed potatoes</skos:prefLabel>
+        <skos:prefLabel xml:lang="de">Kartoffelpüree</skos:prefLabel>
+        <skos:inScheme rdf:resource="http://example.com/dishes"/>
+        <skos:broader rdf:resource="http://example.com/dishes#potatoBased"/>
+    </skos:Concept>
+</rdf:RDF>
